Move StrictMonotonousCounter to module
